modified ft_puthex_pf in order to work on long numbers and modified docs

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: hecmarti <hecmarti@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/19 15:10:15 by hecmarti          #+#    #+#             */
-/*   Updated: 2024/02/19 15:10:28 by hecmarti         ###   ########.fr       */
+/*   Updated: 2024/03/26 10:38:32 by hecmarti         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,7 +24,7 @@ void	ft_putchar_pf(char c, size_t *counter);
 void	ft_putstr_pf(char *str, size_t *counter);
 void	ft_putnbr_pf(int num, size_t *counter);
 void	ft_putuint_pf(unsigned int num, size_t *counter);
-void	ft_puthex_pf(unsigned int num, size_t *counter, char *base);
+void	ft_puthex_pf(unsigned long num, size_t *counter, char *base);
 void	ft_putptr_pf(void *ptr, size_t *counter);
 char	*ft_aux_pf(unsigned long long n, char *base);
